Realize IndexedRegionSet: index-native IntervalSetOps, source-free MCO, trait cleanup
- gtars-core: optimized index-friendly closest() (bidirectional expanding
  scan, fixes +/-2 window bug); IntervalSetOps now 6 methods incl closest;
  gaps() relocated to RegionSet inherent method; add Region::distance_to.
- gtars-overlaprs: Overlapper::iter() (AIList+Bits); MultiChromOverlapper is
  source-free <u32,()> (to_region_set/reduced_by_chr); real index-native
  IntervalSetOps (per-chr sweep) replacing delegation stubs; query API
  unified on min_overlap; IndexedRegionSet is the sole source-owning facade;
  delete regionset_ops.rs / RegionSetOverlaps.
- gtars-genomicdist: delete interval_ranges.rs (~3085 lines) + IntervalRanges;
  relocate StrandedRegionSet ops + RegionSetListOps/pairwise_jaccard.
- downstream (cli/python/wasm/r): subtract->setdiff, overlap queries via
  IndexedRegionSet, intersect_all->IntervalSetOps::intersect.

gtars-py builds (legacy trait removal confirmed non-breaking). Implements
gtars_realize_indexed_regionset_plan_v1; per ADR regionset_source_ownership.
